+++ crates/navidrome/src/repo/genre.rs
@@ -69,6 +69,7 @@ pub async fn get_songs_by_genre(
             user_uploads.r2_key,
             user_uploads.mime_type,
             user_uploads.file_size,
+            user_uploads.sample_rate,
             (SELECT at2.album_id FROM album_tracks at2 WHERE at2.track_id = tracks.xata_id LIMIT 1) AS album_id,
             (SELECT at3.artist_id FROM artist_tracks at3 WHERE at3.track_id = tracks.xata_id LIMIT 1) AS artist_id
         FROM tracks